Move function for external api calls to asyncHelpers
